fix --same-ip-only messing with the alive time
When --same-ip-only is set, only notifications made on the same IP as
the client will be sent back, this is done by checking the IP of the
req, Object, and the notification Object, the latter being set with
another req Object through the POST /new-notification request.

However, as we don't want to give the client the IP of the notification,
we have to delete it, however we not only deleted it from the list of
notifications made on the same IP, but also the original list of
notifications, removing all IP information, and making it impossible to
check the IP. Therefore after the first GET /notifications request it'd
"clear" the notification, and it would mean if there are multiple
clients only one of them would get it, and if that client isn't on the
same IP, they wouldn't even get it, they'd just clear it.

We fix this by merging {} with the notification Object, thereby avoiding
just referencing the original Object, and we can therefore change each
Object independently.